requirements.txt

discord.py
beautifulsoup4
